% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aacf.R
\name{aacf}
\alias{aacf}
\title{Estimate the areal autocorrelation function.}
\usage{
aacf(x)
}
\arguments{
\item{x}{An n x n raster object.}
}
\value{
A list containing matrix and image representations
  of the AACF. Both matrix and image values are normalized
  so that the maximum is equal to 1.
}
\description{
Calculates the areal autocorrelation function (AACF) as the
inverse of the Fourier power spectrum. \code{aacf(x)} returns
the AACF in both matrix and image format.
}
\examples{
# import raster image
data(orforest)

# calculate aacf img and matrix
aacf_list <- aacf(orforest)

# plot resulting aacf image
plot(aacf_list[[2]])
}
