% Generated by roxygen2: do not edit by hand
\name{adaptive_metropolis}
\alias{adaptive_metropolis}
\title{Adaptive Metropolis Sampler}
\usage{
adaptive_metropolis(target_log_density, init = c(0, 0), n_iter = 3000, t0 = 100)
}
\arguments{
\item{target_log_density}{Target log-density.}
\item{init}{Initial state vector.}
\item{n_iter}{Iterations.}
\item{t0}{Adaptation start iteration.}
}
\value{
Matrix of sampled parameter states.
}
\description{
Applies Haario Adaptive Metropolis sampling with dynamic covariance adaptation.
}