% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/adjustBaseline-method.R
\name{adjustBaseline}
\alias{adjustBaseline}
\alias{adjustBaseline,SRM-method}
\title{Adjust Baseline}
\usage{
adjustBaseline(object, hwm)

\S4method{adjustBaseline}{SRM}(object, hwm = 6)
}
\arguments{
\item{object}{a SRM object}

\item{hwm}{a numeric value for window half width for local medians}
}
\value{
a SRM object
}
\description{
Adjust Baseline
}
