% Generated by roxygen2: do not edit by hand
\name{antithetic_mc}
\alias{antithetic_mc}
\title{Antithetic Variates Integrator}
\usage{
antithetic_mc(g, n = 10000)
}
\arguments{
\item{g}{Monotonic integrand function.}
\item{n}{Sample size.}
}
\value{
List containing estimate, standard error, and variance reduction.
}
\description{
Applies antithetic variates negative correlation pairing to reduce integration variance.
}