% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/bpi-method.R
\docType{methods}
\name{bpi}
\alias{bpi}
\alias{bpi,SRM-method}
\title{bpi}
\usage{
bpi(object)

\S4method{bpi}{SRM}(object)
}
\arguments{
\item{object}{a SRM object}
}
\value{
a numeric value for the whole sample base peak intensity (bpi)
}
\description{
bpi
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
