% Generated by roxygen2: do not edit by hand
\name{cftp_perfect_sampler}
\alias{cftp_perfect_sampler}
\title{Coupling From The Past (CFTP) Sampler}
\usage{
cftp_perfect_sampler(max_lookback = 100)
}
\arguments{
\item{max_lookback}{Maximum backward horizon.}
}
\value{
List with exact sample state and lookback time.
}
\description{
Performs exact un-biased MCMC sampling using Coupling From The Past.
}