% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/chromTransform-method.R
\name{chromTransform}
\alias{chromTransform}
\alias{chromTransform,SRM-method}
\title{Transform Chromatograms}
\usage{
chromTransform(object, cf)

\S4method{chromTransform}{SRM}(object, cf)
}
\arguments{
\item{object}{a SRM object}

\item{cf}{matrix, a coefficient matrix generated by MsCoreUtils::coefMA or MsCoreUtils::coefWMA}
}
\value{
a SRM object
}
\description{
Transform Chromatograms
}
