% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineTransitions-method.R
\docType{methods}
\name{combineTransitions,SRM-method}
\alias{combineTransitions,SRM-method}
\title{Combine Transitions}
\usage{
\S4method{combineTransitions}{SRM}(object)
}
\description{
While each parent \code{m/z} will usually have three separat Q3 product ions, SRM transitions are store separately
in the \code{mzML} file. Each index is a single transition of parent \code{m/z} to one of the specified Q3 \code{m/z}.
This method simply combins all Q3 product ion chromatograms into one, to produce a total ion chromatogram based on
the parent \code{m/z} and all Q3 product ions assigned to that parent \code{m/z}
}

