% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{combineTransitions}
\alias{combineTransitions}
\title{Combine Transitions}
\usage{
combineTransitions(object)
}
\arguments{
\item{object}{a SRM object}
}
\value{
a list of SRM total ion chromatograms
}
\description{
Combine Transitions
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

