% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvParams.R
\name{cvParams}
\alias{cvParams}
\title{Extract mzML cvParams}
\usage{
cvParams(xmlDoc)
}
\arguments{
\item{xmlDoc}{a \code{xml} document}
}
\value{
a \code{data.frame} of all avaialble \code{cvParams}. Accession, name and value are all given for each \code{cvParam}
}
\description{
Extract all the \code{cvParams} in the specified \code{mzML} file.
}
\examples{
\dontrun{
library(xml2)
xmlDoc <- read_xml("example_file.mzML")
xmlParams <- cvParams(xmlDoc)
head(xmlParms)
accession                                      name                                    value
1 MS:1001473 selected reaction monitoring chromatogram
2 MS:1000768                    Thermo nativeID format
3 MS:1000563                         Thermo RAW format
4 MS:1000569                                     SHA-1 42823de9154df2488f6bb5f1e5bdbbb84195718d
5 MS:1000492          Thermo Electron instrument model
6 MS:1000529                  instrument serial number                                 TQU01681
}
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
\keyword{internal}

