% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodePeaks.R
\name{decodePeaks}
\alias{decodePeaks}
\title{Decocde peaks}
\usage{
decodePeaks(x, compression, size)
}
\arguments{
\item{compression}{compresison type, "none" or "gzip"}

\item{size}{a numeric value for the number of bytes per element in the byte stream}

\item{xmlDoc}{a \code{base64} encoded vector}
}
\value{
a numeric vector
}
\description{
Decode a \code{binaryDataArray} into a vector of either time or intensity
}
\examples{
\dontrun{
library(xml2)
xmlDoc <- read_xml("example_file.mzML")
xmlPeaks <- binaryArrays(xmlDoc)
intensity <- decodePeaks(xmlPeaks$value[4],size = 8)
head(intensity)
[1] 3.516104 3.427163 3.493258 3.397190 3.430998 3.053472
}
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
\keyword{internal}
