% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectInstrumentModel.R
\name{detectInstrumentModel}
\alias{detectInstrumentModel}
\title{Detect Instrument Model}
\usage{
detectInstrumentModel(mzml_file)
}
\arguments{
\item{mzml_file}{the absolute file path of a valid \code{.mzML} file}
}
\value{
a character string of the MS cvParam description for the instrument model
}
\description{
Detect Instrument Model
}
