% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/detectPeaks-method.R
\name{detectPeaks}
\alias{detectPeaks}
\alias{detectPeaks,SRM-method}
\title{Detect Peaks}
\usage{
detectPeaks(object, raw = TRUE, snthresh, fwhm)

\S4method{detectPeaks}{SRM}(object, raw = TRUE, snthresh = 10, fwhm = 30)
}
\arguments{
\item{object}{a SRM object}
}
\value{
a SRM object
}
\description{
Detect Peaks
}
