% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{estimate_noise}
\alias{estimate_noise}
\title{Estimate Noise (Maxima's)}
\usage{
estimate_noise(x)
}
\arguments{
\item{x}{a numeric vector of intensity}
}
\value{
a numeric value
}
\description{
Make an approximate estimation of baseline noise by taking the mean intensity of all local minima's after removing regions
where a chromatographic peak may be present (local maxima's)
}
\keyword{internal}
