% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/export_to_skyline-method.R
\docType{methods}
\name{export_to_skyline}
\alias{export_to_skyline}
\alias{export_to_skyline,SRM-method}
\title{export_to_skyline}
\usage{
export_to_skyline(object, transitions = NULL)

\S4method{export_to_skyline}{SRM}(object, transitions = NULL)
}
\arguments{
\item{object}{a SRM object}

\item{transitions}{an optional \code{data.frame} containing the following columns for selected (or all) transitions.}
}
\value{
a \code{tibble} of transitions in a format which can be easily used with \code{Skyline}
}
\description{
export_to_skyline
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
