% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/filterPeaks-method.R
\name{filterPeaks}
\alias{filterPeaks}
\alias{filterPeaks,SRM-method}
\title{Filter Peaks}
\usage{
filterPeaks(object)

\S4method{filterPeaks}{SRM}(object)
}
\arguments{
\item{object}{a SRM object}
}
\value{
a \code{SRM} object
}
\description{
Filter Peaks
}
