% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{format_scan_header}
\alias{format_scan_header}
\title{Format scan header}
\usage{
format_scan_header(x)
}
\arguments{
\item{x}{a \code{tibble}}
}
\value{
a vector of tidy scan filters
}
\description{
Format scan header
}
\keyword{internal}
