% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeScales.R
\name{freeScales}
\alias{freeScales}
\title{Free Scales}
\usage{
freeScales(ggplot)
}
\arguments{
\item{ggplot}{a \code{ggplot} object}
}
\value{
a \code{ggplot} object
}
\description{
Allow \code{free} scales in facet plots
}
