% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/get-method.R
\docType{methods}
\name{get}
\alias{get}
\alias{get,SRM-method}
\title{get}
\usage{
get(object, Q1, Q3, method)

\S4method{get}{SRM}(object, Q1, Q3, method)
}
\arguments{
\item{object}{a SRM object}

\item{Q1}{a numeric value for parent m/z}

\item{Q3}{a numeric value for product m/z}

\item{method}{either 'tic' or 'bpi' to return total ion count or base peak intensity respectively}
}
\value{
a \code{tibble} or parent and product mass, and either tic or bpi
}
\description{
get
}
