% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{get_Qmz}
\alias{get_Qmz}
\title{Get Parent (Q1) and Product (Q3) mass values}
\usage{
get_Qmz(x)
}
\arguments{
\item{x}{a vector of scan filters}
}
\value{
a \code{data.frame} of Q1 and Q3
}
\description{
Parse Q1 and Q3 mass values from scan filter string. This function is only used internally in \code{get_scan_header}
}
\keyword{internal}
