% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scan_header.R
\name{get_scan_header}
\alias{get_scan_header}
\title{Extract scan header information}
\usage{
get_scan_header(x)
}
\arguments{
\item{x}{a valid \code{.mzML} file}
}
\value{
a \code{tibble} containing
\itemize{
    \item{\code{header}}
    \item{\code{polarity}}
    \item{\code{Q1}}
    \item{\code{Q3}}
    \item{\code{tidy_head}}
}
}
\description{
Extract key values (parent m/z, product m/z and polarity) from the \code{chromatogram} blocks of an \code{.mzML} file
}
