% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{get_sha1}
\alias{get_sha1}
\title{Extract the SHA-1 value}
\usage{
get_sha1(x)
}
\arguments{
\item{x}{\ valid \code{.mzML} file}
}
\value{
a character string of the SHA-1
}
\description{
Extract the SHA-1 hash value for a \code{.mzML} file
}
\keyword{internal}
