% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQMZs.R
\name{getQMZs}
\alias{getQMZs}
\title{Extract Q1/Q3 masses}
\usage{
getQMZs(x)
}
\arguments{
\item{x}{a \code{data.frame} of \code{cvParams}}
}
\value{
a \code{data.frame} parent and product masses for each SRM transition
}
\description{
Extract parent (Q1) and product (Q3) masses from a \code{data.frame} of \code{cvParams}
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
\keyword{internal}
