% Generated by roxygen2: do not edit by hand
\name{gibbs_sampler}
\alias{gibbs_sampler}
\title{Bivariate Normal Gibbs Sampler}
\usage{
gibbs_sampler(n_iter = 5000, rho = 0.8)
}
\arguments{
\item{n_iter}{Total iterations.}
\item{rho}{Correlation parameter.}
}
\value{
Matrix of sampled (X, Y) states.
}
\description{
Simulates bivariate normal draws using Gibbs sampling full conditionals.
}