% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/groupPeaks-method.R
\name{groupPeaks}
\alias{groupPeaks}
\alias{groupPeaks,SRM-method}
\title{Group Peaks}
\usage{
groupPeaks(object, rt_tolerance)

\S4method{groupPeaks}{SRM}(object, rt_tolerance)
}
\arguments{
\item{object}{a SRM object}

\item{rt_tolerance}{a numeric value for the tolerate retention grouping}
}
\value{
a \code{SRM} object
}
\description{
Group detected peaks based on retention time. Based on \code{MsCoreUtils::group}
}
