% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{groupPeaksInternal}
\alias{groupPeaksInternal}
\title{Group Peaks}
\usage{
groupPeaksInternal(x, rt_tolerance = 0)
}
\arguments{
\item{x}{a numeric vector of retention times}

\item{rt_tolerance}{a numeric value for the tolerate retention grouping}
}
\value{
a numeric vector of retention time groups
}
\description{
Group detected peaks based on retention time. Based on \code{MsCoreUtils::group}
}
\keyword{internal}
