% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/groupSummary-method.R
\name{groupSummary}
\alias{groupSummary}
\alias{groupSummary,SRM-method}
\title{Group Summary}
\usage{
groupSummary(object)

\S4method{groupSummary}{SRM}(object)
}
\arguments{
\item{object}{a \code{SRM} object}
}
\value{
a \code{tibble} of group summary
}
\description{
Create a summary table for peak groups
}
