% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accessor-method.R
\name{header}
\alias{header}
\alias{header,SRM-method}
\title{header}
\usage{
header(object)

\S4method{header}{SRM}(object)
}
\arguments{
\item{object}{a \code{SRM} S4 object}
}
\description{
List object header information
}
