% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idRefs.R
\name{idRefs}
\alias{idRefs}
\title{Extract mzML idRefs}
\usage{
idRefs(xmlDoc)
}
\arguments{
\item{xmlDoc}{a \code{xml} document}
}
\value{
a character vector of all available \code{idRefs}
}
\description{
Extract all the \code{idRefs} in the specified \code{mzML} file. The \code{idRefs} will correspond to the individual scan events. There will be a single \code{idRef} for each
SRM transition. The first \code{idRef} is assigned to the total ion count (TIC) chromatogram of the file.
}
\examples{
\dontrun{
library(xml2)
xmlDoc <- read_xml("example_file.mzML")
xmlRefs <- idRefs(xmlDoc)
head(xmlRefs)
[1] "TIC"                     "SRM SIC 153.01,65.271"
[3] "SRM SIC 153.01,67.232"   "SRM SIC 153.01,109.094"
[5] "SRM SIC 179.022,107.007" "SRM SIC 179.022,134.006"
 }
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
\keyword{internal}

