% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/indexRT-method.R
\name{indexRT}
\alias{indexRT}
\alias{indexRT,SRM-method}
\title{Find Target Retention Time (Rt) for a transition}
\usage{
indexRT(object, index, sampleName)

\S4method{indexRT}{SRM}(object, index, sampleName)
}
\arguments{
\item{object}{a SRM object}

\item{index}{a numeric value for the transition index}

\item{sampleName}{a character string}
}
\value{
a numeric vector or Rt targets
}
\description{
Find Target Retention Time (Rt) for a transition
}
