% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{inegrate_peak_pracma}
\alias{inegrate_peak_pracma}
\title{Integrate Peak (pracma)}
\usage{
inegrate_peak_pracma(rt, int, range)
}
\arguments{
\item{rt}{a numeric vector of retention time}

\item{int}{a numeric vector of intensity}

\item{range}{a numeric vector of peak indicies}
}
\description{
Integrate Peak (pracma)
}
\keyword{internal}
