% Generated by roxygen2: do not edit by hand
\name{inv_transform_exp}
\alias{inv_transform_exp}
\title{Inverse-Transform Exponential Generator}
\usage{
inv_transform_exp(n, rate = 1)
}
\arguments{
\item{n}{Sample size.}
\item{rate}{Rate parameter lambda > 0.}
}
\value{
Numeric vector of exponential random variates.
}
\description{
Generates exponential random variables using inverse-CDF transform X = -log(1-U)/lambda.
}