% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/keepTransitions-method.R
\name{keepTransitions}
\alias{keepTransitions}
\alias{keepTransitions,SRM-method}
\title{Keep Transitions}
\usage{
keepTransitions(object, index_keep)

\S4method{keepTransitions}{SRM}(object, index_keep = c())
}
\arguments{
\item{object}{a SRM object}

\item{index_keep}{a numeric vector of transition index to retain}
}
\value{
a \code{SRM} object
}
\description{
Keep Transitions
}
