% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{local_max}
\alias{local_max}
\title{Local Maxima}
\usage{
local_max(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a vector of local maxima indicies
}
\description{
Find all local maxima's in a time series vector
}
\keyword{internal}
