% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{local_min}
\alias{local_min}
\title{Local Minima}
\usage{
local_min(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a vector of local minima indicies
}
\description{
Find all local minima's in a time series vector
}
\keyword{internal}
