% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchedFilter.R
\name{matchedFilter}
\alias{matchedFilter}
\title{Wraper for xcms matchedFilter peak detection}
\usage{
matchedFilter(rt, int, snthresh)
}
\arguments{
\item{rt}{a numeric vector of chromatogram retention time in minutes}

\item{int}{a numeric vector of chromatogram intensity}

\item{snthresh}{a numeric value for the Signal-to-Noise cut off to use during peak detection}
}
\value{
a \code{tibble} of detected peaks
}
\description{
Wraper for xcms matchedFilter peak detection
}
