% Generated by roxygen2: do not edit by hand
\name{mc_integrate}
\alias{mc_integrate}
\title{Raw Monte Carlo Integration}
\usage{
mc_integrate(g, n = 10000, a = 0, b = 1)
}
\arguments{
\item{g}{Target integrand function.}
\item{n}{Sample size.}
\item{a}{Lower bound.}
\item{b}{Upper bound.}
}
\value{
List containing estimate, standard error, and 95% confidence interval.
}
\description{
Estimates integrals using uniform Monte Carlo sampling.
}