% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accessor-method.R
\name{meta}
\alias{meta}
\alias{meta,SRM-method}
\title{meta}
\usage{
meta(object)

\S4method{meta}{SRM}(object)
}
\arguments{
\item{object}{a \code{SRM} S4 object}
}
\description{
List sample meta data
}
