% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openPRM.R
\name{openPRM}
\alias{openPRM}
\title{Open PRM Files}
\usage{
openPRM(files)
}
\arguments{
\item{files}{a character vector of absolute file paths of PRM files in \code{.mzML} format}
}
\value{
a PRM object
}
\description{
Open and parse PRM files into an \code{S4} PRM Object
}
