% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openSRM.R
\name{openSRM}
\alias{openSRM}
\title{Open SRM Files}
\usage{
openSRM(files, source_type)
}
\arguments{
\item{files}{a character vector of absolute file paths of SRM files in \code{.mzML} format}

\item{source_type}{a character string of the original file format (\code{raw} or \code{lcd})}
}
\value{
an SRM object
}
\description{
Open and parse SRM files into an \code{S4} SRM Object
}
