% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openSRMfile.R
\name{openSRMfile}
\alias{openSRMfile}
\title{Open a SRM-MS file}
\usage{
openSRMfile(mzMLFile)
}
\arguments{
\item{mzMLFile}{a valid \code{.mzML} file}
}
\value{
a \code{SRM} object
 (see \code{\link{SRM-class}})
}
\description{
Open and parse \code{mzML} file into a \code{SRM} object
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
