% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orforest}
\alias{orforest}
\title{NDVI for a portion of southwestern Oregon.}
\format{A raster image with 371 x 371 pixels, with the WGS84
  projection at ~30 meter resolution.
\describe{
  \item{value}{NDVI, range = 0-1, bounds = -123, -122.9, 43.0002, 43.1}
  ...
}}
\usage{
orforest
}
\description{
A raster image of Normalized Difference Vegetation Index (NDVI)
derived from Landsat data for a small portion of SW Oregon state.
}
\keyword{datasets}
