% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accessor-method.R
\name{peaks}
\alias{peaks}
\alias{peaks,SRM-method}
\title{peaks}
\usage{
peaks(object)

\S4method{peaks}{SRM}(object)
}
\arguments{
\item{object}{a \code{SRM} S4 object}
}
\description{
List detected peaks
}
