% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.transition.R
\name{plot.transition}
\alias{plot.transition}
\title{Plot transitions}
\usage{
\method{plot}{transition}(x, n)
}
\arguments{
\item{x}{a \code{transition} class}

\item{n}{the number of the transition (index) to plot}
}
\description{
Plot the SRM-MS TIC and all indivvidual Q3 traces for a selected parent scans
}

