% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAll-method.R, R/plotAll.R
\name{plotAll}
\alias{plotAll}
\alias{plotAll.transition}
\title{Plot all}
\usage{
plotAll(x, n)

\method{plotAll}{transition}(x, n)
}
\arguments{
\item{x}{a \code{transition} class}

\item{n}{the number of the transition (index) to plot}
}
\description{
Plot the SRM-MS TIC and all individual Q3 traces for a selected parent scans'
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
