% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotCompareSample-method.R
\docType{methods}
\name{plotCompareSample}
\alias{plotCompareSample}
\alias{plotCompareSample,SRM-method}
\title{Plot Compare Sample}
\usage{
plotCompareSample(object, index, sampleName)

\S4method{plotCompareSample}{SRM}(object, index, sampleName)
}
\arguments{
\item{object}{a SRM object}

\item{index}{a numeric value of the transition index to plot}

\item{sampleName}{a character vector of \code{sampleName} to plot}
}
\value{
a ggplot plot object
}
\description{
Plot Compare Sample
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
