% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotGroup-method.R
\name{plotGroup}
\alias{plotGroup}
\alias{plotGroup,SRM-method}
\title{Plot Group}
\usage{
plotGroup(object, group)

\S4method{plotGroup}{SRM}(object, group)
}
\arguments{
\item{object}{a SRM object}

\item{group}{a character string of the \code{GroupID}}
}
\value{
a ggplot plot object
}
\description{
Plot Group
}
