% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotMulti-method.R
\docType{methods}
\name{plotMulti}
\alias{plotMulti}
\alias{plotMulti,SRM-method}
\title{plotMulti}
\usage{
plotMulti(object, idn)

\S4method{plotMulti}{SRM}(object, idn)
}
\arguments{
\item{object}{a SRM object}

\item{idn}{the index number of a transition to plot}
}
\description{
plotMulti
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

