% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotParent-method.R
\name{plotParent}
\alias{plotParent}
\alias{plotParent,SRM-method}
\title{Plot Parent}
\usage{
plotParent(object, parentMass)

\S4method{plotParent}{SRM}(object, parentMass)
}
\arguments{
\item{object}{a SRM object}

\item{parentMass}{a numeric value of the parent mass (Q1) to extract}
}
\value{
a ggplot plot object
}
\description{
Plot Parent
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
