% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotPeakArea-method.R
\name{plotPeakArea}
\alias{plotPeakArea}
\alias{plotPeakArea,SRM-method}
\title{Plot Peak Area}
\usage{
plotPeakArea(object, index, sampleName)

\S4method{plotPeakArea}{SRM}(object, index, sampleName)
}
\arguments{
\item{object}{a SRM object}

\item{index}{a numeric value of the transition index to plot}

\item{sampleName}{a character string of \code{sampleName} to plot}
}
\value{
a ggplot plot object
}
\description{
Plot Peak Area
}
