% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotSample-method.R
\name{plotSample}
\alias{plotSample}
\alias{plotSample,SRM-method}
\title{Plot Sample}
\usage{
plotSample(object, sampleName, polarity)

\S4method{plotSample}{SRM}(object, sampleName, polarity)
}
\arguments{
\item{object}{a SRM object}

\item{sampleName}{a character string of a valid \code{sampleName}}

\item{polarity}{a character string of either \code{pos} or \code{neg} for positive and negative ionisation mode respectively}
}
\value{
a ggplot plot object
}
\description{
Plot Sample
}
