% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotSRM-method.R
\docType{methods}
\name{plotSRM}
\alias{plotSRM}
\alias{plotSRM,SRM-method}
\title{Plot SRM}
\usage{
plotSRM(object, idn)

\S4method{plotSRM}{SRM}(object, idn)
}
\arguments{
\item{object}{a SRM objecy}

\item{idn}{the index number of a transition to plot}
}
\description{
Plot SRM
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

