% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{plotTransform}
\alias{plotTransform}
\title{Plot Transform}
\usage{
plotTransform(object, index, sampleID)
}
\arguments{
\item{object}{a SRM object}

\item{index}{a numeric value of the transition index to plot}

\item{sampleID}{a character string of a valid \code{sampleName}}
}
\value{
a ggplot plot object
}
\description{
Plot Transform
}
