% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pracma_peak_picking.R
\name{pracma_peak_picking}
\alias{pracma_peak_picking}
\title{Pracma Peak-Picking}
\usage{
pracma_peak_picking(rt, int)
}
\arguments{
\item{rt}{a numeric vector of retention time}

\item{int}{a numeric vector of intensity}
}
\value{
a `tibble` of peak infomation (rt, rtmin, rtmax, int, area, peakId)
}
\description{
A wrapper around `pracma::findPeaks` for peak detection and integration
}
