% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm_methods.R
\name{print.srm}
\alias{print.srm}
\title{Print an srm object}
\usage{
\method{print}{srm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{srm}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an srm object
}
\examples{
mat = matrix(c(0, 1, 2, 2, 0, 1, 1, 2, 0), 3, 3)
rownames(mat) = colnames(mat) = c("A", "B", "C")
fit = srm(mat)
print(fit)

}
