% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm_methods.R
\name{print.summary.srm}
\alias{print.summary.srm}
\title{Print a summary.srm object}
\usage{
\method{print}{summary.srm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.srm}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a summary.srm object
}
