% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/removeSample-method.R
\name{removeSample}
\alias{removeSample}
\alias{removeSample,SRM-method}
\title{Remove Sample}
\usage{
removeSample(object, sampleName)

\S4method{removeSample}{SRM}(object, sampleName)
}
\arguments{
\item{object}{a SRM object}

\item{sampleName}{a character string of a \code{sampleName} to remove}
}
\value{
a \code{SRM} object
}
\description{
Remove Sample
}
