% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/removeTransitions-method.R
\name{removeTransitions}
\alias{removeTransitions}
\alias{removeTransitions,SRM-method}
\title{Keep Transitions}
\usage{
removeTransitions(object, index_out)

\S4method{removeTransitions}{SRM}(object, index_out = c())
}
\arguments{
\item{object}{a SRM object}

\item{index_keep}{a numeric vector of transition index to remove}
}
\value{
a \code{SRM} object
}
\description{
Keep Transitions
}
