% Generated by roxygen2: do not edit by hand
\name{robbins_monro}
\alias{robbins_monro}
\title{Robbins-Monro Stochastic Approximation}
\usage{
robbins_monro(noisy_g, init = 0, n_iter = 1000, c = 1)
}
\arguments{
\item{noisy_g}{Noisy measurement function.}
\item{init}{Initial guess.}
\item{n_iter}{Iterations.}
\item{c}{Step decay constant.}
}
\value{
Sequence of theta estimates.
}
\description{
Performs stochastic root-finding using the Robbins-Monro algorithm.
}