% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savitzky_golay_opt.R
\name{savitzky_golay_opt}
\alias{savitzky_golay_opt}
\title{(Optimised) Savitzky Golay}
\usage{
savitzky_golay_opt(rt, int, ford)
}
\arguments{
\item{rt}{a numeric vector of retention time}

\item{int}{a numeric vector of intensity}

\item{ford}{a numeric value for the filter order (2 = quadratic filter, 4 = quartic)}
}
\value{
a numeric vector of smoothed intensity
}
\description{
(Optimised) Savitzky Golay
}
