% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanPolarity.R
\name{scanPolarity}
\alias{scanPolarity}
\title{Extract scan polarity}
\usage{
scanPolarity(x)
}
\arguments{
\item{x}{a \code{data.frame} of \code{cvParams}}
}
\value{
a vector of scan polarities
}
\description{
Extract the scan polarity from a \code{data.frame} of \code{cvParams}
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
\keyword{internal}
