% Generated by roxygen2: do not edit by hand
\name{score_function_gradient}
\alias{score_function_gradient}
\title{Score Function Gradient Estimator}
\usage{
score_function_gradient(h, theta = 1, n = 5000)
}
\arguments{
\item{h}{Payoff/reward function.}
\item{theta}{Parameter theta.}
\item{n}{Sample size.}
}
\value{
Estimated expectation gradient d/d_theta E[h(X)].
}
\description{
Estimates expectation gradients using the log-derivative score function method.
}