% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\docType{methods}
\name{show,SRM-method}
\alias{show,SRM-method}
\title{show-sRm}
\usage{
\S4method{show}{SRM}(object)
}
\arguments{
\item{object}{a SRM object}
}
\description{
show-sRm
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

