% Generated by roxygen2: do not edit by hand
\name{sim_poisson_process}
\alias{sim_poisson_process}
\title{Poisson Process Simulator}
\usage{
sim_poisson_process(T_max = 10, rate = 2)
}
\arguments{
\item{T_max}{Simulation time horizon.}
\item{rate}{Event arrival rate lambda > 0.}
}
\value{
Vector of event arrival times.
}
\description{
Simulates homogeneous Poisson point process arrival times.
}