% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{SRM-class}
\alias{SRM-class}
\title{SRM}
\description{
A S4 class to store selective reaction monitoring (SRM) mass spectrometry (MS) data
}
\section{Slots}{

\describe{
\item{\code{header}}{}
}}

