%% File Name: srm-package.Rd
%% File Version: 0.05

\name{srm-package}
\alias{srm-package}
\alias{srm}
\docType{package}

\title{
\packageTitle{srm}
}

\description{
\packageDescription{srm}
}

\author{
\packageAuthor{srm}

Maintainer: \packageMaintainer{srm}
}

%% \code{Sys.time}: \tab syssys \cr

%\details{
% xx
%}

%Statistical inference is conducted for multiply imputed datasets.
%Inference for nested multiple imputation (Rubin, 2003) is also included.



%----
%For the following functions, an own function maintainer is defined
%which should be contacted instead of the package maintainer:
%
%Alexander Robitzsch <a.robitzsch@bifie.at> \cr
%\code{\link{BIFIE.logistreg}}
%----

\references{
Bond, C. F., & Malloy, T. E. (2018). Social relations analysis of dyadic data structure:
The general case. In T. E. Malloy. 
\emph{Social Relations Modeling of Behavior in Dyads and Groups} (Ch. 14).
Academic Press. \doi{10.1016/B978-0-12-811967-9.00014-X}

Gill, P. S., & Swartz, T. B. (2001). Statistical analyses for round robin interaction 
data. \emph{Canadian Journal of Statistics, 29}(2), 321-331.
\doi{10.2307/3316080}

Nestler, S. (2018). Likelihood estimation of the multivariate social relations model. 
\emph{Journal of Educational and Behavioral Statistics, 43}(4), 387-406.
\doi{10.3102/1076998617741106}

Warner, R. M., Kenny, D. A., & Soto, M. (1979). A new round robin analysis of variance for
social interaction data. \emph{Journal of Personality and Social Psychology, 37}, 
1742-1757. \doi{10.1037/0022-3514.37.10.1742} 
}

\keyword{package}

\seealso{
See also the \R packages \pkg{amen} and \pkg{TripleR} for estimating the
social relations model.
}

%% \examples{
%% ##   |-----------------------------------------------------------------
%% }
