% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm_effects.R
\name{srm_effects_long}
\alias{srm_effects_long}
\title{This function calculates actor, partner, and unique effects via srm framework}
\usage{
srm_effects_long(x, type = c("actor", "partner", "unique"), time = NULL)
}
\arguments{
\item{x}{A matrix or list of matrices (for longitudinal data)}

\item{type}{A character string specifying the type of effect to calculate.}

\item{time}{For longitudinal data (list of matrices), specify which time points to analyze}
}
\value{
A matrix or list of matrices with actor/partner/unique effects
}
\description{
This function calculates actor, partner, and unique effects via srm framework
}
