% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm_plots.R
\name{srm_plotactor}
\alias{srm_plotactor}
\title{plot actor effect
function plots actor and partner effect from srm_effects(actor/partner)}
\usage{
srm_plotactor(mat, n = 10, type = c("actor", "partner"))
}
\arguments{
\item{mat}{matrix from srm_effects(actor/partner).}

\item{n}{number of top actors/partners to plot.}

\item{type}{A string specifying the type of effect to plot.}
}
\value{
A ggplot object.
}
\description{
plot actor effect
function plots actor and partner effect from srm_effects(actor/partner)
}
