% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm_plots.R
\name{srm_plotdyadic}
\alias{srm_plotdyadic}
\title{plot dyadic effect}
\usage{
srm_plotdyadic(mat)
}
\arguments{
\item{mat}{matrix from srm_effects(unique).}
}
\value{
A ggplot heatmap object.
}
\description{
This function creates a heatmap of dyadic effects from srm_effects(unique)?.
It colors negative values in red and positive values in blue
}
