% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localsurface.R
\name{ssc}
\alias{ssc}
\title{Mean summit curvature.}
\usage{
ssc(x)
}
\arguments{
\item{x}{An n x n raster object.}
}
\value{
A numeric value representing the average curvature of
  surface peaks.
}
\description{
Calculates the mean summit curvature of a raster. Mean summit
curvature is the average principle curvature of local maximas
on the surface.
}
\examples{
# import raster image
data(normforest)

# calculate mean summit curvature
Ssc <- ssc(normforest)
}
