% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/tic-method.R
\docType{methods}
\name{tic}
\alias{tic}
\alias{tic,SRM-method}
\title{tic}
\usage{
tic(object)

\S4method{tic}{SRM}(object)
}
\arguments{
\item{object}{a SRM object}
}
\value{
a numeric value for the whole sample total ion count (tic)
}
\description{
tic
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
