% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trade_net}
\alias{trade_net}
\title{Simulated trade network (longitudinal)}
\format{
A named list of three 10x10 numeric matrices. Countries:
USA, CHN, DEU, JPN, GBR, FRA, KOR, IND, BRA, CAN.
}
\usage{
data(trade_net)
}
\description{
A named list of three 10x10 matrices representing simulated bilateral
trade intensity between 10 countries across three time periods
(2015, 2017, 2019). Useful for demonstrating longitudinal SRM analysis.
}
\examples{
data(trade_net)
fit = srm(trade_net)
summary(fit)
}
\keyword{datasets}
