% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accessor-method.R
\name{transitions}
\alias{transitions}
\alias{transitions,SRM-method}
\title{Accessor Methods for the SRM S4 class}
\usage{
transitions(object)

\S4method{transitions}{SRM}(object)
}
\arguments{
\item{object}{a \code{SRM} S4 object}
}
\description{
List all transitions
}
