% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_idrefs.R
\name{unique_idrefs}
\alias{unique_idrefs}
\title{Parse idRef}
\usage{
unique_idrefs(object)
}
\arguments{
\item{object}{a \code{SRM} object}
}
\value{
a character vector of parsed \code{idRefs}
}
\description{
Parse \code{idRefs} when dynamic acqusition has been used. Where dynamic acquistion is used, each SRM transition
contains at retention window aswell as a pair of parent (Q1) and product (Q3) masses. \code{unique_idrefs} parses the
\code{idRefs} to extract all information needed to identify a unique SRM-MS transition
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
\keyword{internal}
