% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryArrays.R
\name{binaryArrays}
\alias{binaryArrays}
\title{Extract mzML binary arrays}
\usage{
binaryArrays(xmlDoc)
}
\arguments{
\item{xmlDoc}{a \code{xml} document}
}
\value{
a list of three elements; array name, array value and array precision. Array name is character vector specifying
the type of array (time array or intensity). Array value is a large character vector where each element
is the \code{base64} encdoded value of the \code{binaryDataArray}. Arrray precision is the value specifing the encoded precision
of each array (32-bit or 64-bit)
}
\description{
Extract all the \code{binaryDataArrays} in the specified \code{mzML} file.
}
\examples{
\dontrun{
library(xml2)
xmlDoc <- read_xml("example_file.mzML")
xmlPeaks <- binaryArrays(xmlDoc)
names(xmlPeaks)
[1] "name"  "value"
 }
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
\keyword{internal}

