% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{detectPeaks}
\alias{detectPeaks}
\title{Detect Peaks}
\usage{
detectPeaks(object, method, ...)
}
\arguments{
\item{object}{a SRM object}

\item{method}{a character string indicating the peak detection method to use.
*
*
*}

\item{...}{any additional parameters that are required by the peak detection method}

\item{raw;}{logical}
}
\value{
a SRM object
}
\description{
Detect Peaks
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
