% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/detectPeaks-method.R
\name{detectPeaks}
\alias{detectPeaks}
\alias{detectPeaks,SRM-method}
\title{Detect Peaks}
\usage{
detectPeaks(object, snthresh, peakwidth)

\S4method{detectPeaks}{SRM}(object, snthresh, peakwidth)
}
\arguments{
\item{object}{a SRM object}

\item{snthresh}{a numeric value for the signal-to-noise threshold to use}

\item{peakwidth}{a numeric vector indicating the minimum and maximum tolerated peak width}
}
\value{
a SRM object
}
\description{
Detect Peaks
}
