% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/detectPeaks-method.R
\docType{methods}
\name{detectPeaks}
\alias{detectPeaks}
\alias{detectPeaks,SRM-method}
\title{Detect Peaks}
\usage{
detectPeaks(object, method, ...)

\S4method{detectPeaks}{SRM}(object, method, ...)
}
\arguments{
\item{object}{a SRM object}

\item{method}{a character string indicating the peak detection method to use.
*
*
*}

\item{...}{any additional parameters that are required by the peak detection method}
}
\value{
a SRM object
}
\description{
Detect Peaks
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
