% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Extract meta information from .mzML file}
\usage{
get_meta(x, type = "raw")
}
\arguments{
\item{x}{a valid \code{.mzML} file}

\item{type}{a character string of the original file format (\code{raw} or \code{lcd})}
}
\value{
a \code{tibble} containing;
\itemize{
\item mzML Schema
\item Acquisition Date
\item Acquisition Time
\item Instrument Model
\item File ID
}
}
\description{
Extract a minimal amount meta data from a \code{.mzML} file
}
