% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/meta-method.R
\docType{methods}
\name{meta}
\alias{meta}
\alias{meta,SRM-method}
\title{Show meta data}
\usage{
meta(object)

\S4method{meta}{SRM}(object)
}
\arguments{
\item{object}{a SRM object}
}
\description{
Show meta data

Show meta data
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

