% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openSRMfile.R
\name{openSRMfile}
\alias{openSRMfile}
\title{Open a SRM-MS file}
\usage{
openSRMfile(filename)
}
\arguments{
\item{filename}{a \code{mzML} file}
}
\value{
a \code{SRM} object
 (see \code{\link{SRM-class}})
}
\description{
Function for reading and parsing \code{mzML} file containing SRM-MS data
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

