% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/peakAsymmetry-method.R
\docType{methods}
\name{peakAsymmetry}
\alias{peakAsymmetry}
\alias{peakAsymmetry,SRM-method}
\title{Peak Asymmetry}
\usage{
peakAsymmetry(object)

\S4method{peakAsymmetry}{SRM}(object)
}
\arguments{
\item{object}{a SRM object}
}
\value{
a \code{SRM} object
}
\description{
Peak Asymmetry
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
