% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotMulti-method.R
\docType{methods}
\name{plotMulti}
\alias{plotMulti}
\alias{plotMulti,SRM-method}
\title{plotMulti}
\usage{
plotMulti(object, idn, addLabels = FALSE, labels = NULL)

\S4method{plotMulti}{SRM}(object, idn, addLabels = TRUE, labels = NULL)
}
\arguments{
\item{object}{a SRM object}

\item{idn}{the index number of a transition to plot}

\item{addLabels}{logical; If \code{TRUE} then labels for each \code{geom_line} are added}

\item{labels}{an optional character vector of label names. Default is \code{NULL}. If \code{addLabels} is \code{TRUE} and \code{labels} is \code{NULL} then \code{object@index} is used for label text.}
}
\description{
plotMulti
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
