% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotMulti-method.R
\docType{methods}
\name{plotMulti}
\alias{plotMulti}
\alias{plotMulti,SRM-method}
\title{plotMulti}
\usage{
plotMulti(object, idn, labels = NULL)

\S4method{plotMulti}{SRM}(object, idn, labels = NULL)
}
\arguments{
\item{object}{a SRM object}

\item{idn}{the index number of a transition to plot}

\item{labels}{an optional character vector of label names. Default is \code{NULL}}
}
\description{
plotMulti
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

