% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotSRM-method.R
\name{plotSRM}
\alias{plotSRM}
\alias{plotSRM,SRM-method}
\title{Plot SRM}
\usage{
plotSRM(object, index, type = "overlay")

\S4method{plotSRM}{SRM}(object, index, type = "overlay")
}
\arguments{
\item{object}{a SRM object}

\item{index}{a numeric value of the transition index to plot}

\item{type}{a character string of either \code{overlay} or \code{facet}}
}
\value{
a ggplot plot object
}
\description{
Plot SRM
}
