% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotTransform-method.R
\name{plotTransform}
\alias{plotTransform}
\alias{plotTransform,SRM-method}
\title{Plot Transform}
\usage{
plotTransform(object, index_n, sampleID)

\S4method{plotTransform}{SRM}(object, index_n, sampleID)
}
\arguments{
\item{object}{a SRM object}

\item{index_n}{a numeric value of the transition index to plot}

\item{sampleID}{a character string of a valid \code{sampleName}}
}
\value{
a ggplot plot object
}
\description{
Plot Transform
}
