% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/smoothChrom-method.R
\name{smoothChrom}
\alias{smoothChrom}
\alias{smoothChrom,SRM-method}
\title{Smooth Chromatograms}
\usage{
smoothChrom(object, cf)

\S4method{smoothChrom}{SRM}(object, cf)
}
\arguments{
\item{object}{a SRM object}

\item{cf}{matrix, a coefficient matrix generated by MsCoreUtils::coefMA or MsCoreUtils::coefWMA}
}
\value{
a SRM object
}
\description{
Smooth Chromatograms
}
