% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{SRM-class}
\alias{SRM-class}
\title{SRM}
\description{
A S4 class to store selective reaction monitoring (SRM) mass spectrometry (MS) data
}
\section{Slots}{

\describe{
\item{\code{env}}{class environment}

\item{\code{SHA1}}{character vector of the file SHA1 checksum value}

\item{\code{meta}}{list containing file meta data}

\item{\code{totalIonCount}}{\code{data.frame} of sample total ion count (TIC)}

\item{\code{index}}{character vector of scan index's; ie \code{SRM SIC 341.014,114.098}}

\item{\code{peaks}}{list of peak data}

\item{\code{header}}{\code{data.frame} of summary information for each scan index; parentMz, Q3mz, totalIonCount, basePeakInt}
}}

