% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm_effects.R
\name{srm_effects}
\alias{srm_effects}
\title{This function calculates actor, partner,and unique effects via srm framework}
\usage{
srm_effects(x, type = c("actor", "partner", "unique"))
}
\arguments{
\item{x}{can be a df or matrix. if df- converted to matrix by netify}

\item{type}{A character string specifying the type of effect to calculate.}
}
\value{
A vector of actor/partner/unique effects
}
\description{
This function calculates actor, partner,and unique effects via srm framework
}
