% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm_effects.R
\name{srm_stats}
\alias{srm_stats}
\title{function for means, variance}
\usage{
srm_stats(
  x,
  type = c("rowmeans", "colmeans", "totalmeans", "actor_variance",
    "dyadic/unique_variance", "partner_variance", "dyadic/relationship_covariance",
    "actor_partner_covariance")
)
}
\arguments{
\item{x}{A square matrix}

\item{type}{A character string specifying the type of effect to calculate.}
}
\value{
A vector of actor/partner/unique effects
}
\description{
function for means, variance
}
