% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/transitions-method.R
\docType{methods}
\name{transitions}
\alias{transitions}
\alias{transitions,SRM-method}
\title{Show transitions}
\usage{
transitions(object)

\S4method{transitions}{SRM}(object)
}
\arguments{
\item{object}{a SRM object}
}
\description{
Show transitions
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
